Fixed test flakiness by separating integration tests.
