+++ full-backend-tests/src/test/java/org/graylog/plugins/views/StartPageRecentActivityIT.java
@@ -16,60 +16,26 @@
  */
 package org.graylog.plugins.views;
 
+import org.graylog.testing.completebackend.Lifecycle;
 import org.graylog.testing.completebackend.apis.GraylogApis;
 import org.graylog.testing.completebackend.apis.Streams;
 import org.graylog.testing.completebackend.apis.Users;
 import org.graylog.testing.containermatrix.annotations.ContainerMatrixTest;
 import org.graylog.testing.containermatrix.annotations.ContainerMatrixTestsConfiguration;
 import org.graylog2.plugin.streams.StreamRuleType;
-import org.graylog2.shared.security.RestPermissions;
-import org.junit.jupiter.api.BeforeAll;
 
-import java.util.Collections;
-import java.util.List;
 import java.util.Map;
 
 import static org.hamcrest.core.StringContains.containsString;
 
-@ContainerMatrixTestsConfiguration
-public class StartPageIT {
+@ContainerMatrixTestsConfiguration(serverLifecycle = Lifecycle.CLASS)
+public class StartPageRecentActivityIT {
+
     private final GraylogApis api;
 
-    private static final Users.User user = new Users.User(
-            "john.doe2",
-            "asdfgh",
-            "John",
-            "Doe",
-            "john.doe2@example.com",
-            false,
-            30_000,
-            "Europe/Vienna",
-            Collections.emptyList(),
-            List.of(RestPermissions.DASHBOARDS_CREATE)
-    );
-
-    public StartPageIT(GraylogApis api) {
+    public StartPageRecentActivityIT(GraylogApis api) {
         this.api = api;
     }
-
-    @BeforeAll
-    public void init() {
-        api.users().createUser(this.user);
-    }
-
-    @ContainerMatrixTest
-    void testCreateLastOpenedItem() {
-        api.postWithResource("/views/search", user,"org/graylog/plugins/views/startpage-save-search-request.json", 201);
-        api.postWithResource("/views", user,"org/graylog/plugins/views/startpage-views-request.json", 200);
-
-        var validatableResponse = api.get("/views", user, Map.of(), 200);
-        var id = validatableResponse.extract().jsonPath().get("views[0].id").toString();
-
-        api.get("/views/" + id, user, Map.of(), 200);
-        validatableResponse = api.get("/startpage/lastOpened", user, Map.of(), 200);
-        validatableResponse.assertThat().body("lastOpened[0].grn", containsString(id));
-    }
-
 
     @ContainerMatrixTest
     void testCreateRecentActivity() {
